+++ R/reorder_genes.R
@@ -18,6 +18,7 @@
 #'
 #' @return ClusterSet-class object
 #' @export reorder_genes
+#' @importFrom Matrix Matrix
 #' @keywords internal
 setGeneric("reorder_genes", 
            function(object=NULL, 
@@ -99,10 +100,9 @@ setMethod("reorder_genes",
     for (cur_clust in names(object@gene_clusters)) {
       
     gene_cluster_names <- object@gene_clusters[[cur_clust]]
-    expression_matrix <- object@data[gene_cluster_names,]
-    
+
     set.seed(123)
-    hclust_res <- amap::hcluster(expression_matrix,
+    hclust_res <- amap::hcluster(object@data[gene_cluster_names,],
                                  method = method,
                                  diag = FALSE,
                                  upper = FALSE,
@@ -120,7 +120,7 @@ setMethod("reorder_genes",
   }
   
   # Reorder genes in data slot
-  object@data <- object@data[unlist(object@gene_clusters, use.names = FALSE),]
+  object@data <- Matrix::Matrix(object@data[unlist(object@gene_clusters, use.names = FALSE),], sparse=TRUE)
   
   return(object)
 })